requirements.txt
streamlit==1.25.0  
python-dotenv==1.0.0
pypdf==3.15.1
langchain==0.93.0 
langchain-huggingface==0.0.6
langchain-ollama==0.0.3
langchain-community==0.0.1
faiss-cpu==1.7.4.post2  
